new constraint id: all normal constraint has id

+++ src/protocol/src/pr_model.rs
@@ -7,7 +7,7 @@ pub struct PrParticle {
 }
 
 pub struct PrConstraint {
-	pub id: isize,
+	pub id: i32,
 	pub particles: Vec<usize>,
 }
